+++ actionmailer/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Reset `ActionMailer::Base.deliveries` after every test in
+    `ActionDispatch::IntegrationTest`.
+
+    *Yves Senn*
+
+
 ## Rails 5.0.0.beta2 (February 01, 2016) ##
 
 *   No changes.